requirements.txt
langchain==0.1.16
langchain-community
langchain-openai==0.0.8
fastapi
langserve
sse_starlette
python-dotenv
ipykernel
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
pymupdf
arxiv
wikipedia
validators==0.28.1
huggingface_hub
streamlit
duckduckgo-search